Add video playback!
Woohoo, everything works! So far... This modifies the link handler
to be far more readable, because comma-delimited strings suck. It
also adds the basic video player along with some video stats.

Would be nice to hack the player a bit too, so need to see what's
possible with the Player APIs at some point.