Update core and abstract Youtube initialisation code
Created a new repo for the Youtube Wrapper which uses Laravel 4.1,
so updating this one to match (else conflicts). Starting off by
abstracting the loading of the Google and Youtube services.
This reduces route code from 11 lines to 4, and completely removes
another dozen from the top of the file. Still loads to do though!