Ajaxify channel playlists
Removes playlists code from the ChannelController and adds a new
method to the PlaylistController instead. This is then called by
Ajax and loaded into the appropriate tab.

Currently this acts on click, which IMHO is shit. This needs
to happen automatically or it's pointless. Probably do some
work on the wrapper package to support the calls there directly.